refactor: extract composable functions for screen sections
