+++ dukan_presentation/src/commonMain/kotlin/net/thechance/mena/dukan/presentation/screen/createDukan/content/CreateDukanContentSelectLocation.kt
@@ -19,7 +19,7 @@ import mena.dukan_presentation.generated.resources.location
 import net.thechance.mena.designsystem.presentation.component.text.Text
 import net.thechance.mena.designsystem.presentation.component.textField.TextField
 import net.thechance.mena.designsystem.presentation.theme.theme.Theme
-import net.thechance.mena.dukan.presentation.screen.createDukan.content.component.Map
+import net.thechance.mena.dukan.presentation.screen.createDukan.component.Map
 import net.thechance.mena.dukan.presentation.viewModel.createDukan.CreateDukanInteractionListener
 import net.thechance.mena.dukan.presentation.viewModel.createDukan.CreateDukanUiState
 import org.jetbrains.compose.resources.painterResource
@@ -35,57 +35,66 @@ fun CreateDukanContentSelectLocation(
             .fillMaxSize()
             .padding(horizontal = Theme.spacing._16)
     ) {
-        item {
-            Text(
-                text = stringResource(Res.string.dukan_location),
-                style = Theme.typography.title.medium,
-                color = Theme.colorScheme.shadePrimary
-            )
-        }
-        item {
-            Text(
-                text = stringResource(Res.string.enter_address_and_pick_dukan_location_on_map),
-                style = Theme.typography.body.small,
-                color = Theme.colorScheme.shadeSecondary
-            )
-        }
-        item {
-            Text(
-                modifier = Modifier.padding(top = Theme.spacing._16, bottom = Theme.spacing._4),
-                text = stringResource(Res.string.location),
-                style = Theme.typography.title.small,
-                color = Theme.colorScheme.shadePrimary
-            )
-        }
-        item {
-            Map(
-                modifier = Modifier
-                    .clip(RoundedCornerShape(Theme.radius.md))
-                    .fillMaxWidth()
-                    .height(244.dp),
-                isLocked = state.isMapLocked,
-                anchorLocation = state.pointerLocation,
-                cameraPosition = state.cameraPosition,
-                onMapClick = listener::onMapClicked,
-                onCameraMoved = listener::onCameraMoved,
-                onEditClick = listener::onEditMapLocationClicked
-            )
-        }
-        item {
-            Text(
-                modifier = Modifier.padding(top = Theme.spacing._12, bottom = Theme.spacing._4),
-                text = stringResource(Res.string.address),
-                style = Theme.typography.title.small,
-                color = Theme.colorScheme.shadePrimary
-            )
-        }
-        item {
-            TextField(
-                value = state.address,
-                onValueChanged = listener::onAddressChanged,
-                hint = "",
-                leadingIcon = painterResource(Res.drawable.ic_store_location),
-            )
-        }
+        item { DukanLocationHeader() }
+        item { DukanLocationMapSection(state, listener) }
+        item { DukanAddressSection(state, listener) }
     }
+}
+
+@Composable
+private fun DukanLocationHeader() {
+    Text(
+        text = stringResource(Res.string.dukan_location),
+        style = Theme.typography.title.medium,
+        color = Theme.colorScheme.shadePrimary
+    )
+    Text(
+        text = stringResource(Res.string.enter_address_and_pick_dukan_location_on_map),
+        style = Theme.typography.body.small,
+        color = Theme.colorScheme.shadeSecondary
+    )
+}
+
+@Composable
+private fun DukanLocationMapSection(
+    state: CreateDukanUiState,
+    listener: CreateDukanInteractionListener
+) {
+    Text(
+        modifier = Modifier.padding(top = Theme.spacing._16, bottom = Theme.spacing._4),
+        text = stringResource(Res.string.location),
+        style = Theme.typography.title.small,
+        color = Theme.colorScheme.shadePrimary
+    )
+    Map(
+        modifier = Modifier
+            .clip(RoundedCornerShape(Theme.radius.md))
+            .fillMaxWidth()
+            .height(244.dp),
+        isLocked = state.isMapLocked,
+        anchorLocation = state.pointerLocation,
+        cameraPosition = state.cameraPosition,
+        onMapClick = listener::onMapClicked,
+        onCameraMoved = listener::onCameraMoved,
+        onEditClick = listener::onEditMapLocationClicked
+    )
+}
+
+@Composable
+private fun DukanAddressSection(
+    state: CreateDukanUiState,
+    listener: CreateDukanInteractionListener
+) {
+    Text(
+        modifier = Modifier.padding(top = Theme.spacing._12, bottom = Theme.spacing._4),
+        text = stringResource(Res.string.address),
+        style = Theme.typography.title.small,
+        color = Theme.colorScheme.shadePrimary
+    )
+    TextField(
+        value = state.address,
+        onValueChanged = listener::onAddressChanged,
+        hint = "",
+        leadingIcon = painterResource(Res.drawable.ic_store_location),
+    )
 }
