+++ dukan_presentation/src/commonMain/kotlin/net/thechance/mena/dukan/presentation/screen/createDukan/content/CreateDukanContentSelectStyle.kt
@@ -17,8 +17,8 @@ import mena.dukan_presentation.generated.resources.style
 import net.thechance.mena.designsystem.presentation.component.text.Text
 import net.thechance.mena.designsystem.presentation.theme.theme.MenaTheme
 import net.thechance.mena.designsystem.presentation.theme.theme.Theme
-import net.thechance.mena.dukan.presentation.screen.createDukan.content.component.dukanstyle.ColorOptionsPlaceholder
-import net.thechance.mena.dukan.presentation.screen.createDukan.content.component.dukanstyle.DukanStyleOptions
+import net.thechance.mena.dukan.presentation.screen.createDukan.component.dukanstyle.ColorOptionsPlaceholder
+import net.thechance.mena.dukan.presentation.screen.createDukan.component.dukanstyle.DukanStyleOptions
 import net.thechance.mena.dukan.presentation.util.stubPreviews.PreviewCreateDukanInteractionListener
 import net.thechance.mena.dukan.presentation.viewModel.createDukan.CreateDukanInteractionListener
 import net.thechance.mena.dukan.presentation.viewModel.createDukan.CreateDukanUiState
@@ -33,50 +33,54 @@ fun CreateDukanContentSelectStyle(
     state: CreateDukanUiState
 ) {
     LazyColumn(
-        modifier = Modifier.fillMaxSize()
+        modifier = Modifier
+            .fillMaxSize()
             .background(Theme.colorScheme.background.surface)
             .padding(horizontal = Theme.spacing._16)
     ) {
-        item {
-            CustomizeDukanHeader()
-        }
-        item {
-            Text(
-                text = stringResource(Res.string.color),
-                style = Theme.typography.title.small,
-                color = Theme.colorScheme.shadePrimary,
-                modifier = Modifier.padding(top = Theme.spacing._16, bottom = Theme.spacing._4)
-            )
-        }
-        item {
-            FlowRow(
-                horizontalArrangement = Arrangement.spacedBy(Theme.spacing._8),
-                verticalArrangement = Arrangement.spacedBy(Theme.spacing._8),
-            ) {
-                state.dukanColors.forEach {
-                    ColorOptionsPlaceholder(
-                        backgroundColor = Color(it.color),
-                        onClick = { listener.onColorClicked(it) },
-                        isSelected = it.id == state.selectedColor?.id
-                    )
-                }
-            }
-        }
-        item {
-            Text(
-                text = stringResource(Res.string.style),
-                style = Theme.typography.title.small,
-                color = Theme.colorScheme.shadePrimary,
-                modifier = Modifier.padding(top = Theme.spacing._16, bottom = Theme.spacing._4)
-            )
-        }
-        item {
-            DukanStyleOptions(
-                listener = listener,
-                state = state
+        item { CustomizeDukanHeader() }
+        item { DukanColorSection(listener, state) }
+        item { DukanStyleSection(listener, state) }
+    }
+}
+
+@Composable
+private fun DukanColorSection(
+    listener: CreateDukanInteractionListener,
+    state: CreateDukanUiState
+) {
+    Text(
+        text = stringResource(Res.string.color),
+        style = Theme.typography.title.small,
+        color = Theme.colorScheme.shadePrimary,
+        modifier = Modifier.padding(top = Theme.spacing._16, bottom = Theme.spacing._4)
+    )
+    FlowRow(
+        horizontalArrangement = Arrangement.spacedBy(Theme.spacing._8),
+        verticalArrangement = Arrangement.spacedBy(Theme.spacing._8),
+    ) {
+        state.dukanColors.forEach {
+            ColorOptionsPlaceholder(
+                backgroundColor = Color(it.color),
+                onClick = { listener.onColorClicked(it) },
+                isSelected = it.id == state.selectedColor?.id
             )
         }
     }
+}
+
+@Composable
+private fun DukanStyleSection(
+    listener: CreateDukanInteractionListener,
+    state: CreateDukanUiState
+) {
+    Text(
+        text = stringResource(Res.string.style),
+        style = Theme.typography.title.small,
+        color = Theme.colorScheme.shadePrimary,
+        modifier = Modifier.padding(top = Theme.spacing._16, bottom = Theme.spacing._4)
+    )
+    DukanStyleOptions(listener = listener, state = state)
 }
 
 @Composable
